requirements.txt
streamlit
sounddevice
numpy
matplotlib
torchaudio
speechbrain
scikit-learn
librosa
